% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiffall.R
\name{setdiffall}
\alias{setdiffall}
\title{Summarize setdiff between two vectors}
\usage{
setdiffall(x, y)
}
\arguments{
\item{x, y}{vectors}
}
\value{
\code{setdiffall} returns a list of 3 elements: setdiff in both directions and the intersection
}
\description{
Given two vectors, apply setdiff in both directions and intersect
and prints a nice summary message of results
}
\examples{
x <- 1:3
y <- 1:4
setdiffall(x, y)


}
